Day - 14 - Route wise permission management

+++ app/Http/Controllers/Backend/User/UsersController.php
@@ -8,6 +8,8 @@ use App\Models\Designation;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Spatie\Permission\Models\Role;
+use Spatie\Permission\Models\Permission;
 
 class UsersController extends Controller
 {
@@ -18,6 +20,10 @@ class UsersController extends Controller
      */
     public function index()
     {
+        if (!request()->user()->hasPermissionTo('user.list')) {
+            abort(401, 'You have not permission to list of the user !');
+        }
+
         $users = User::all();
         return view('backend.pages.users.index', compact('users'));
     }
@@ -29,6 +35,29 @@ class UsersController extends Controller
      */
     public function create()
     {
+        // Create Role
+        // $role = User::createRole('Super Admin');
+
+        // Give Role to Permission
+        // $role->givePermissionTo('user.create');
+        // $role->givePermissionTo('user.list');
+        // $role->givePermissionTo('user.view');
+        // $role->givePermissionTo('user.edit');
+        // $role->givePermissionTo('user.delete');
+
+        // Get authenticated user data
+        // $user = User::find(1);
+        $user = request()->user();
+
+        // Role Assign
+        // $user->assignRole('Super Admin');
+
+
+        // Check if logged in user has "user.create" permission or not
+        if (!request()->user()->hasPermissionTo('user.create')) {
+            abort(401, 'You have not permission to create a user !');
+        }
+
         $designations = Designation::select('id', 'name')->get();
         return view('backend.pages.users.create', compact('designations'));
     }
@@ -41,6 +70,10 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
+        if (!request()->user()->hasPermissionTo('user.create')) {
+            abort(401, 'You have not permission to create a user !');
+        }
+
         // Validate our data
         $request->validate(
             [
@@ -79,6 +112,7 @@ class UsersController extends Controller
         $user->present_address = $request->present_address;
         $user->parmanent_address = $request->parmanent_address;
         $user->status = $request->status;
+        $user->created_by = request()->user()->id;
         $user->save();
 
         if ($request->image) {
@@ -99,6 +133,10 @@ class UsersController extends Controller
      */
     public function show($id)
     {
+        if (!request()->user()->hasPermissionTo('user.view')) {
+            abort(401, 'You have not permission to view a user !');
+        }
+
         //
     }
 
@@ -110,6 +148,10 @@ class UsersController extends Controller
      */
     public function edit($id)
     {
+        if (!request()->user()->hasPermissionTo('user.edit')) {
+            abort(401, 'You have not permission to edit a user !');
+        }
+
         $user = User::find($id);
         if (!is_null($user)) {
             $designations = Designation::select('id', 'name')->get();
@@ -129,6 +171,10 @@ class UsersController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!request()->user()->hasPermissionTo('user.edit')) {
+            abort(401, 'You have not permission to update a user !');
+        }
+
         $user = User::find($id);
 
         if (!is_null($user)) {
@@ -136,9 +182,9 @@ class UsersController extends Controller
             $request->validate(
                 [
                     'name' => 'required|string|max:255|min:2',
-                    'email' => 'required|string|email|max:255|unique:users,email,'.$user->id,
-                    'phone_no' => 'required|max:255|max:15|min:11|unique:users,phone_no,'.$user->id,
-                    'username' => 'required|string|max:20|alpha_num|unique:users,username,'.$user->id,
+                    'email' => 'required|string|email|max:255|unique:users,email,' . $user->id,
+                    'phone_no' => 'required|max:255|max:15|min:11|unique:users,phone_no,' . $user->id,
+                    'username' => 'required|string|max:20|alpha_num|unique:users,username,' . $user->id,
                     'image' => 'nullable|image|max:2048',
                     'present_address' => 'nullable|max:255',
                     'parmanent_address' => 'nullable|max:255',
@@ -169,6 +215,7 @@ class UsersController extends Controller
             $user->present_address = $request->present_address;
             $user->parmanent_address = $request->parmanent_address;
             $user->status = $request->status;
+            $user->updated_by = request()->user()->id;
             $user->save();
 
             if ($request->image) {
@@ -180,7 +227,7 @@ class UsersController extends Controller
             }
 
             session()->flash('success', 'User has been updated !');
-        }else{
+        } else {
             session()->flash('error', 'Sorry ! No User has been found !');
         }
         return redirect()->route('admin.users.index');
@@ -194,13 +241,18 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
+        if (!request()->user()->hasPermissionTo('user.delete')) {
+            abort(401, 'You have not permission to delete a user !');
+        }
+
         $user = User::find($id);
         if (!is_null($user)) {
             // First Delete User Image & then Delete the user
-            UploadHelper::delete('images/users/'.$user->image);
+            UploadHelper::delete('images/users/' . $user->image);
+            $user->deleted_by = request()->user()->id;
             $user->delete();
             session()->flash('success', 'User has been deleted !');
-        }else{
+        } else {
             session()->flash('error', 'Sorry ! No User has been found !');
         }
         return redirect()->route('admin.users.index');
